update demoactor example
