+++ examples/demo_actor/demo_actor/demo_actor_service.py
@@ -2,24 +2,17 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 
-from flask import Flask, jsonify
-from flask_dapr.actor import DaprActor
-
-from dapr.conf import settings
+from fastapi import FastAPI
+from dapr.ext.fastapi import DaprActor
 from demo_actor import DemoActor
 
-app = Flask(f'{DemoActor.__name__}Service')
 
-# Enable DaprActor Flask extension
+app = FastAPI(title=f'{DemoActor.__name__}Service')
+
+# Add Dapr Actor Extension
 actor = DaprActor(app)
-# Register DemoActor
-actor.register_actor(DemoActor)
 
-# This route is optional.
-@app.route('/')
-def index():
-    return jsonify({'status': 'ok'}), 200
-
-
-if __name__ == '__main__':
-    app.run(port=settings.HTTP_APP_PORT)
+@app.on_event("startup")
+async def startup_event():
+    # Register DemoActor
+    await actor.register_actor(DemoActor)
