+++ examples/demo_actor/demo_actor/requirements.txt
@@ -1,2 +1,1 @@
-dapr-dev>=dapr-dev-0.6.0a0.dev67
-flask-dapr-dev>=0.1.0a0.dev67
+dapr-ext-fastapi-dev>=0.6.0a0.dev0
